complete manage profile
