+++ src/components/profile/ManageProfile.js
@@ -1,54 +1,106 @@
+import { LockIcon, EditIcon } from '../../images';
+
 export default function ManageProfile() {
     return (
         <>
             <div className=" h-screen flex justify-center flex-col bg-black">
                 <div className="flex justify-center mt-10">
-                    <h1 className="text-white">Manage Profiles:</h1>
+                    <h1 className="text-white text-4xl">Manage Profiles:</h1>
                 </div>
-                <div className="flex justify-center gap-3  mt-3">
-                    <div>
+                <div className="flex justify-center gap-4 mt-6">
+                    <div className="relative">
                         <img
                             className="w-[100px]"
                             src="http://occ-0-2339-769.1.nflxso.net/dnm/api/v6/K6hjPJd6cR6FpVELC5Pd6ovHRSk/AAAABQ_JEhgVlFamty5bFR0ofXrH6YH0NJ868l21auwAbp78Qx2jeEuB088_MY2tepK6ppRoDQ7Fet7t8c4CI7t1xn_zbkLj0QtJU8Ot.png?r=749"
                             alt="1"
                         />
+                        <div className="fill-white w-[100px] h-[100px] absolute top-0 left-0  bg-black/50 z-30 ">
+                            <div className="absolute top-10 left-10 w-7 h-7 ">
+                                <EditIcon />
+                            </div>
+                        </div>
                         <span className="flex justify-center text-white">
                             Aommei
                         </span>
+                        <div className="w-4 h-4 mt-5 mx-auto">
+                            <LockIcon />
+                        </div>
                     </div>
-                    <div>
+                    <div className="relative">
                         <img
                             className="w-[100px]"
                             src="http://occ-0-2339-769.1.nflxso.net/dnm/api/v6/K6hjPJd6cR6FpVELC5Pd6ovHRSk/AAAABQ_JEhgVlFamty5bFR0ofXrH6YH0NJ868l21auwAbp78Qx2jeEuB088_MY2tepK6ppRoDQ7Fet7t8c4CI7t1xn_zbkLj0QtJU8Ot.png?r=749"
                             alt="1"
                         />
-                        <span className="flex justify-center  text-white">
-                            Soffy
+                        <div className="fill-white w-[100px] h-[100px] absolute top-0 left-0  bg-black/50 z-30 ">
+                            <div className="absolute top-10 left-10 w-7 h-7 ">
+                                <EditIcon />
+                            </div>
+                        </div>
+                        <span className="flex justify-center text-white">
+                            Aommei
                         </span>
+                        <div className="w-4 h-4 mt-5 mx-auto">
+                            <LockIcon />
+                        </div>
                     </div>
-                    <div>
+                    <div className="relative">
                         <img
                             className="w-[100px]"
                             src="http://occ-0-2339-769.1.nflxso.net/dnm/api/v6/K6hjPJd6cR6FpVELC5Pd6ovHRSk/AAAABQ_JEhgVlFamty5bFR0ofXrH6YH0NJ868l21auwAbp78Qx2jeEuB088_MY2tepK6ppRoDQ7Fet7t8c4CI7t1xn_zbkLj0QtJU8Ot.png?r=749"
                             alt="1"
                         />
-                        <span className="flex justify-center  text-white">
-                            Mint
+                        <div className="fill-white w-[100px] h-[100px] absolute top-0 left-0  bg-black/50 z-30 ">
+                            <div className="absolute top-10 left-10 w-7 h-7 ">
+                                <EditIcon />
+                            </div>
+                        </div>
+                        <span className="flex justify-center text-white">
+                            Aommei
                         </span>
+                        <div className="w-4 h-4 mt-5 mx-auto">
+                            <LockIcon />
+                        </div>
                     </div>
-                    <div>
+                    <div className="relative">
                         <img
                             className="w-[100px]"
                             src="http://occ-0-2339-769.1.nflxso.net/dnm/api/v6/K6hjPJd6cR6FpVELC5Pd6ovHRSk/AAAABQ_JEhgVlFamty5bFR0ofXrH6YH0NJ868l21auwAbp78Qx2jeEuB088_MY2tepK6ppRoDQ7Fet7t8c4CI7t1xn_zbkLj0QtJU8Ot.png?r=749"
                             alt="1"
                         />
-                        <span className="flex justify-center  text-white">
-                            Tum
+                        <div className="fill-white w-[100px] h-[100px] absolute top-0 left-0  bg-black/50 z-30 ">
+                            <div className="absolute top-10 left-10 w-7 h-7 ">
+                                <EditIcon />
+                            </div>
+                        </div>
+                        <span className="flex justify-center text-white">
+                            Aommei
                         </span>
+                        <div className="w-4 h-4 mt-5 mx-auto">
+                            <LockIcon />
+                        </div>
+                    </div>
+                    <div className="relative">
+                        <img
+                            className="w-[100px]"
+                            src="http://occ-0-2339-769.1.nflxso.net/dnm/api/v6/K6hjPJd6cR6FpVELC5Pd6ovHRSk/AAAABQ_JEhgVlFamty5bFR0ofXrH6YH0NJ868l21auwAbp78Qx2jeEuB088_MY2tepK6ppRoDQ7Fet7t8c4CI7t1xn_zbkLj0QtJU8Ot.png?r=749"
+                            alt="1"
+                        />
+                        <div className="fill-white w-[100px] h-[100px] absolute top-0 left-0  bg-black/50 z-30 ">
+                            <div className="absolute top-10 left-10 w-7 h-7 ">
+                                <EditIcon />
+                            </div>
+                        </div>
+                        <span className="flex justify-center text-white">
+                            Aommei
+                        </span>
+                        <div className="w-4 h-4 mt-5 mx-auto">
+                            <LockIcon />
+                        </div>
                     </div>
                 </div>
-                <div className="flex justify-center mt-10">
-                    <button className="border-2 border-white bg-white  text-black w-136 h-69 p-3">
+                <div className="flex justify-center mt-14">
+                    <button className="border-2 border-white bg-white  text-black  px-3">
                         Done
                     </button>
                 </div>
