--- src/images/EditIcon.js
@@ -1,11 +1,0 @@
-export function EditIcon() {
-    return (
-        <svg
-            className="text-white"
-            xmlns="http://www.w3.org/2000/svg"
-            viewBox="0 0 32 32"
-        >
-            <path d="M2 26h28v2H2zM25.4 9c.8-.8.8-2 0-2.8l-3.6-3.6c-.8-.8-2-.8-2.8 0l-15 15V24h6.4l15-15zm-5-5L24 7.6l-3 3L17.4 7l3-3zM6 22v-3.6l10-10l3.6 3.6l-10 10H6z" />
-        </svg>
-    );
-}
